+++ trunk/src/LV2/gx_amp_stereo.lv2/gxamp_stereo.cpp
@@ -237,7 +237,7 @@ private:
   bool cab_changed()
     {return abs(cab - clevel_ ) > 0.1;}
   bool buffsize_changed() 
-    {return abs(bufsize - cur_bufsize) != 0;}
+    {return abs(static_cast<int>(bufsize) - static_cast<int>(cur_bufsize)) != 0;}
   void update_cab()
     {cab = (clevel_ ); c_old_model_ = c_model_;}
   bool change_cab()
